Create function to number objects (figures, tables, etc.)

This function creates a function that numbers objects (e.g., figures, tables, etc.) in an R Markdown file.

create_numeration_fun(prefix_start = NULL, prefix_end = NULL,
  start_at = 1L, add_space = TRUE, prefix_bold = TRUE, env = new.env(),
  varName = make.names(paste0(prefix_start, prefix_end)))

create_numeration_fun0(fmt = "%g %s", start_at = 1L, env = new.env(),
  varName = make.names(fmt))

Arguments

prefix_start
Sting that will be used in prefix before count, e.g. "Fig.".

prefix_end
Sting that will be used in prefix after count, e.g. " figue".

start_at
A number (an integer, greater than or equal to 1) at which numbering starts.

add_space
Logical. If TRUE (default), a space between prefix and the main body of caption/description is added.

prefix_bold
Logical. If TRUE (default), double asterisk "**" will be used to enclose the prefix, ant it will be interpreted as bold in R markdown.

env
(environment object) An environment to store the count number.

varName
(string) A string with variable name to be created in environment `env` in which number of counts is stored.

fmt
a character vector of format strings, each of up to 8192 bytes to be passed to function sprintf. fmt must include one tag for a numeric value of object's number, (e.g. %g) and one tag for a string with description (%s). Default is "%g %s". The tag for a number must be before the tag for a string.

Value

A function to number objects and add caption/description. The default result of created function is a string. The default arguments are caption = "" - a string with description, restart = FALSE - logical. If TRUE, numbering is restarted; output = "caption" - A string, that indicated what kind of output should be returned. If output = "caption", a numbered prefix with description is returned, e.g. "Fig.1 Random noise". If output = "count" number of objects that are numbered by this function (after the last restart) is returned. Other arguments are ignored.

Examples

iFig   <- create_numeration_fun("Fig.")

# Alternative syntax:
iFig   <- create_numeration_fun0("**Fig. %g** %s")

lapply(paste("Automatically numbered caption for object",LETTERS[1:5]),
      FUN = iFig);

#> [[1]]
#> [1] "**Fig. 1** Automatically numbered caption for object A"
#> 
#> [[2]]
#> [1] "**Fig. 2** Automatically numbered caption for object B"
#> 
#> [[3]]
#> [1] "**Fig. 3** Automatically numbered caption for object C"
#> 
#> [[4]]
#> [1] "**Fig. 4** Automatically numbered caption for object D"
#> 
#> [[5]]
#> [1] "**Fig. 5** Automatically numbered caption for object E"
#> 

# Continue numering
iFig("The 6-th object.")

#> [1] "**Fig. 6** The 6-th object."

# Count of numbered objects
iFig(output = "count")

#> [1] 6

# Restart numering
iFig("Numbering is restarted.", restart = TRUE)

#> [1] "**Fig. 1** Numbering is restarted."

iFig(output = "count")

#> [1] 1

# Create separate functions to number different types of objects.

iTable   <- create_numeration_fun("Table.")
iFig     <- create_numeration_fun("Fig.")
iSection <- create_numeration_fun("Section ")

iSection(); iTable(); iSection(); iFig(); iSection();

#> [1] "**Section 1**"

#> [1] "**Table.1**"

#> [1] "**Section 2**"

#> [1] "**Fig.1**"

#> [1] "**Section 3**"