Extract and Replace parts of knitrContainer objects

The Methods to extract and replace elements of a knitrContainer object

# S3 method for knitrContainer
"["(x, i, ..., section = NULL, level = 1,
  drop = FALSE)

# S3 method for knitrContainer
"[["(x, i, ..., drop = TRUE)

# S3 method for knitrContainer
"$"(x, name)

Arguments

x
a knitrContainer Object

i
index of object inside knitrContainer

...
ignored

drop
For [[: drop unnecessary dimensions, see drop and Extract. Ignored for [, as otherwise invalid knitrContainer objects might result.

name
name of the data column to extract.

j
selecting ...

l
selecting ...

name
name of the data column to extract.

Details

These Methods can be used for selecting/deleting, and extracting or setting elements of a knitrContainer object. Extracting: [, [[, and $.

The version with single square brackets ([) returns the resulting knitrContainer object. [[ yields corresponding knitrContainer element returned. $ returns the selected element. Shortcuts. Three shortcuts to conveniently extract much needed parts of the object are defined: x[[]] is equivalent to print_all. x$. returns ... x$.. returns ... Replacing: [<-, [[<-, and $<-. value gives the values to be assigned. [[<- replaces parts of the ... [<- replaces parts of the ... $<- replaces a ... $..<- is again an abbreviation for the ...