Print and summary methods for knitrContainer

Print summary of knitrContainer object.

# S3 method for knitrContainer
print(x, n = 30, preview = 25, units = "Kb", ...)

Arguments

x
knitrContainer object.

n
The number of rows (objects) to display. Default n = 30.

preview
length of text (number of characters) to be previewed in summary. Default is 25.

units
the units to be used in formatting and printing the size. Allowed values for the different standards are

standard = "legacy":

"b", "Kb", "Mb", "Gb", "Tb", "Pb", "B", "KB", "MB", "GB", "TB" and "PB".

standard = "IEC":

"B", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB", "ZiB" and "YiB".

standard = "SI":

"B", "kB", "MB", "GB", "TB", "PB", "EB", "ZB" and "YB".

For all standards, unit = "auto" is also allowed. If standard = "auto", any of the "legacy" and units are allowed. See ‘Formatting and printing object sizes’ for details.

...
arguments to be passed to or from other methods.

See also

Other knitrContainer functions: Join, add_as_heading, add_as_, as.knitrContainer, knitrContainer-class, print_all, summary

Examples

# Find examples in link `knitrContainer-class`