Collect and Print Multiple Objects in a Knitr R Markdown Report

The purpose of knitrContainer is to collect objects (especially ones generated in a loop) and print them in a knitr/rmarkdown report.

WHY? Some objects such as pander tables and plotly plots, are not printed from inside a loop and are not displayed in `knitr` reports. `knitrContainer` solves this problem by providing convenience functions to get these objects printed and included in HTML files.

Objects

What objects can be collectd and printed with knitrContainer?

An object used by knitrContainer can be any object, that is includible in an R list and that is printable, such as:

objects printable with pander (pander: An R Pandoc Writer, pander);

ggplot2 (ggplot2 is a plotting system, ggplot);

plotly (Plotly R Library, plot_ly);

text, strings;

etc.

References

Some ideas for functions in knitrContainer package are taken from this answer on github.com.

See also