[!] Print objects from knitrContainer

Function is designed to be used in a knitr/rmarkdown file which generates HTML output.
The function takes every element (i.e., object) of a knitrContainer, prints it appropriately: either as text (using cat), as htmlwidget (using cat) or object as-is using print. Then attaches HTML dependencies for htmlwidgets to show them correctly.

print_all(container, env = parent.frame(), widget_as_html = "auto", ...)

extract_and_print(container, env = parent.frame(), ...)

print_objects(container, env = parent.frame(), ...)

Arguments

container
knitr_container object.

env
Environment in which evaluation of expressions an assignments (objects added with add_as_cmd and add_as_data) take place.

widget_as_html
(TURE, FALSE, "auto") Should htmlwidget be printed as an HTML code or as an object (default is "auto").

...
not used.

Details

Functions print_objects() and extract_and_print() are deprecated.

See also

Examples

# Find examples in link `knitrContainer-class`